[CASSH Server] v1.7.0 : Cluster mode on

+++ tests/launch_demo_server.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 
+CARRY_POSTGRES=false
 ENTRYPOINT=
 MOUNT_VOL=
+PORT=8080
 
 function usage()
 {
     echo "Usage:"
-    echo "$0 [-d|--debug] [-h|---help] [-s|--server_code_path <filepath>]"
+    echo "$0 [-d|--debug] [-h|---help] [-p|--port <port>] [-s|--server_code_path <filepath>]"
     echo ""
     exit 0
 }
@@ -27,6 +29,11 @@ case $key in
     shift # past argument
     shift # past value
     ;;
+    -p|--port)
+    PORT="$2"
+    shift # past argument
+    shift # past value
+    ;;
     -d|--debug)
     ENTRYPOINT='--entrypoint /bin/bash'
     shift # past argument
@@ -44,18 +51,24 @@ set -- "${POSITIONAL[@]}" # restore positional parameters
 
 echo 'INIT : LBC - SSH'
 
-echo 'Starting Postgresql server'
-docker run --rm -p 5432:5432 --name demo-postgres -e POSTGRES_PASSWORD=mysecretpassword postgres:latest &
+if [ ! "$(docker ps -q -f name=demo-postgres)" ]; then
+    CARRY_POSTGRES=true
 
-sleep 10
+    echo 'Starting Postgresql server'
+    docker run --rm -p 5432:5432 --name demo-postgres -e POSTGRES_PASSWORD=mysecretpassword postgres:latest &
 
-echo "Initialize server database"
-python tests/init_pg.py
+    sleep 10
 
-sleep 5
+    echo "Initialize server database"
+    python tests/init_pg.py
+
+    sleep 5
+fi
 
 echo 'Starting CA-SSH demo server'
-docker run -it --rm -p 8080:8080 ${MOUNT_VOL} ${ENTRYPOINT} nbeguier/cassh-server:latest
+docker run -it --rm -p "${PORT}":8080 ${MOUNT_VOL} ${ENTRYPOINT} nbeguier/cassh-server:latest
 
-echo 'Stoping Postgresql server'
-docker stop demo-postgres
+if $CARRY_POSTGRES; then
+    echo 'Stoping Postgresql server'
+    docker stop demo-postgres
+fi
